requirements.txt
streamlit==1.13.0
pandas==1.5.1
pandas-profiling==3.4.0
pillow==9.2.0
pycaret==2.3.10
streamlit-pandas-profiling==0.1.3